Add start/end dates to enable filtering for MCA
To make date-based filtering work for MCA customers, the API requires
two additional query string parameters; see [1] for more details.

Unfortunately, the OpenAPI spec [2] used to generate the client does
not include the required `startDate` and `endDate`.

So we decided to fork the https://github.com/Azure/azure-sdk-for-go
repo and add the missing params. This is a temporary measure.

For the long-term, we're getting in touch with Microsoft [3] to
include the missing params in the spec, so future releases will not
require any patch.

[1]: https://docs.microsoft.com/en-us/azure/cost-management-billing/costs/manage-automation#get-usage-details-for-a-scope-during-specific-date-range
[2]: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/consumption/resource-manager/Microsoft.Consumption/stable/2019-10-01
[3]: https://github.com/Azure/azure-rest-api-specs/pull/19499